+++ src/hydrogen/module/session/keycloak.clj
@@ -11,17 +11,28 @@
    {:keycloak
     {:realm (env/env '["KEYCLOAK_REALM" Str])
      :url (env/env '["KEYCLOAK_URL" Str])
-     :clientId (env/env '["KEYCLOAK_CLIENT_ID" Str])}}})
+     :clientId (env/env '["KEYCLOAK_CLIENT_ID" Str])}}
+
+   :magnet.buddy-auth/jwt-oidc
+   {:claims
+    {:iss (env/env '["OIDC_ISSUER_URL" Str])
+     :aud (env/env '["OIDC_AUDIENCE" Str])}
+    :jwks-uri (env/env '["OIDC_JWKS_URI" Str])}
+
+   :duct.middleware.buddy/authentication
+   {:backend :token
+    :token-name "Bearer"
+    :authfn (ig/ref :magnet.buddy-auth/jwt-oidc)}})
 
 (defn- session-config [options project-ns]
   (cond->
    (session-config-base project-ns)
 
     (:add-example-api? options)
-    (assoc (keyword (str project-ns ".api/example")) {})))
+    (assoc (keyword (str project-ns ".api/example"))
+           {:auth-middleware (ig/ref :duct.middleware.buddy/authentication)})))
 
 (defmethod ig/init-key :hydrogen.module/session.keycloak [_ options]
   (fn [config]
     (let [project-ns (project-ns config options)]
       (core/merge-configs config (session-config options project-ns)))))
-