+++ CHANGELOG.md
@@ -3,9 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## UNRELEASED
+### Added
+- Config necessary for backend token verification
+
 ## [0.1.0] - 2019-06-10
 
 ### Added
 - Initial version!
 
 [0.1.0]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/releases/tag/v0.1.0
+[Unreleased]: https://github.com/magnetcoop/hydrogen.module.session.keycloak/compare/v0.1.0...HEAD/